+++ src/components/DropdownCourse/styles.ts
@@ -5,7 +5,7 @@ import { theme } from "../../global/styles/theme";
 
 export const styles = StyleSheet.create({
   container: {
-    width: 331,
+    width: 360,
     height: 60,
     flexDirection: 'row',
     alignItems: 'center',
@@ -44,11 +44,12 @@ export const styles = StyleSheet.create({
     backgroundColor: theme.colors.menuInputBackground,
     borderWidth: 1,
     borderColor: theme.colors.menuInputBorder,
-    width: 331,
+    width: 360,
     height: 48,
     alignItems: 'center',
     marginBottom: 15,
     borderRadius: 10,
+    top:50
   },
   dropdown1BtnTxtStyle: { fontFamily: theme.fonts.text400, color: theme.colors.menuInputBorder, textAlign: 'left', fontSize: 16 },
   dropdown1DropdownStyle: { backgroundColor: theme.colors.menuInputBackground },
