--- src/components/ModalPerfilAluno/index.tsx
@@ -1,56 +1,0 @@
-import React from 'react';
-import {KeyboardAvoidingView, TouchableOpacity, FlatList} from 'react-native';
-
-import {ButtonSave} from '../ButtonSave/index';
-import {EditInput} from '../EditInput/index';
-import {AvatarEdit} from '../AvatarEdit/index';
-// import {EditButton} from '../EditButton/index';
-import {EditField} from '../EditField/index';
-import {LinearGradient} from "expo-linear-gradient";
-import { AreaInteresse } from '../AreaInteresse';
-// import {RectButtonProps} from "react-native-gesture-handler";
-
-import {themes} from './styles';
-import {theme} from "../../global/styles/theme";
-
-export function EditPerfilAluno() {
-
-    const data = [
-        {'key':'Matematica'},
-        {'key':'Desenvolvimento Web'},
-        {'key':'Vadiar'}
-    ]
-
-    const {primary, secondary} = theme.colorFilter;
-
-    return (
-        <KeyboardAvoidingView style={themes.centered}>
-            <LinearGradient colors={[primary, secondary]}>
-                <KeyboardAvoidingView style={themes.modal}>
-
-                    <AvatarEdit urlImage={'https://github.com/neom200.png'}/>
-
-                    <EditInput placeholder="Nome completo"/>
-                    <EditInput placeholder="Aluno"/>
-                    <EditInput placeholder="Campus"/>
-                    <EditInput placeholder="Curso"/>
-
-                    <EditField placeholder="Areas de interesse"/>
-
-                    <FlatList
-                        style={themes.list}
-                        data={data}
-                        renderItem={({item}) => <AreaInteresse dado={item['key']}/>}
-                    />
-
-                    <TouchableOpacity onPress={() => {
-                        alert("DADOS ALTERADOS COM SUCESSO")
-                    }} style={{marginTop: 250}}>
-                        <ButtonSave name="Salvar"/>
-                    </TouchableOpacity>
-
-                </KeyboardAvoidingView>
-            </LinearGradient>
-        </KeyboardAvoidingView>
-    )
-}
