+++ src/routes/index.tsx
@@ -15,8 +15,8 @@ import { PerfilProfessor } from '../screens/PerfilProfessor';
 import { Atalhos } from '../screens/Atalhos';
 import { Home } from "../screens/Home";
 import { Details } from '../screens/Details';
-// import { EditPerfilAluno } from '../components/ModalPerfilAluno'
-import { EditPerfilProfessor } from '../components/ModalPerfilProfessor'
+import { EditPerfilProfessor } from '../screens/EditPerfilProfessor'
+import { EditPerfilAluno } from '../screens/EditPerfilAluno'
 
 
 import { Feather } from '@expo/vector-icons';
