+++ src/screens/Atalhos/index.tsx
@@ -8,7 +8,7 @@ import { themes } from './styles';
 import { Background } from '../../components/Background';
 import { Profile } from "../../components/Profile";
 import { EditPerfilProfessor } from '../../components/ModalPerfilProfessor';
-import { EditPerfilAluno } from '../../components/ModalPerfilAluno';
+import { EditPerfilAluno } from '../EditPerfilAluno';
 
 import { BottomSheet, BottomSheetRef } from 'react-native-sheet';
 import { useNavigation } from '@react-navigation/native';
