+++ src/screens/EditPerfilAluno/styles.ts
@@ -1,0 +1,25 @@
+import { StyleSheet } from "react-native";
+//import {theme} from "../../global/styles/theme";
+
+export const themes = StyleSheet.create({
+    container: {
+        flex: 1,
+        left:20,
+    },
+    container2: {
+        flex: 1,
+        left:20,
+    },
+    container3: {
+        flex: 7,
+        left:20,
+    },
+    list: {
+        flex:1,
+        top:50,
+        alignContent:'center',
+        borderWidth:1,
+        width:360,
+        height:150,
+    }
+})
